LUM-49: Moved benchmark tests to io.lumify.benchmark package. Renamed all integration test classes "*IT.java" and configured Maven Failsafe plugin to run integration tests in the integration-test phase of the build instead of using Surefire to run them as unit tests. Configured Maven profiles so any combination of unit and integration tests can be run or skipped for a particular build.
