+++ integration-test/src/test/java/io/lumify/benchmark/BenchmarkCategory.java
@@ -1,4 +1,4 @@
-package io.lumify.it;
+package io.lumify.benchmark;
 
 /**
  * Use this with {@link org.junit.experimental.categories.Category annotations} on tests that are measuring performance.
