+++ integration-test/src/test/java/io/lumify/benchmark/VertexBenchmark.java
@@ -1,6 +1,8 @@
-package io.lumify.it;
+package io.lumify.benchmark;
 
 import com.google.common.collect.ImmutableList;
+import io.lumify.it.TestClassAndMethod;
+import io.lumify.it.VertextTestBase;
 import io.lumify.web.clientapi.LumifyApi;
 import io.lumify.web.clientapi.codegen.ApiException;
 import io.lumify.web.clientapi.codegen.VertexApi;
@@ -20,7 +22,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThat;
 
 @Category(BenchmarkCategory.class)
-public class VertexBenchmarkTest extends VertextTestBase {
+public class VertexBenchmark extends VertextTestBase {
     private TimedExecution timedExecution;
 
     @Rule
