+++ integration-test/src/test/java/io/lumify/it/VertexIT.java
@@ -14,7 +14,7 @@ import java.util.Map;
 
 import static org.junit.Assert.*;
 
-public class VertexIntegrationTest extends VertextTestBase {
+public class VertexIT extends VertextTestBase {
 
     private static final String PROPERTY_QUERY_STRING = PROPERTY_VALUE_PREFIX;
     private static final String NO_MATCHING_PROPERTY_VALUE = "NoMatchingProperty";
